pyproject.toml
[tool.poetry]
name = "pycon-2023"
version = "1.0.0"
description = "You can find everything that related to the speech in here."
authors = ["reqww <my.gurbanov@gmail.com>"]
readme = "README.md"
packages = [{include = "pycon_2023"}]

[tool.poetry.dependencies]
# You could face some problems with pytype. You should play around python version.
python = "^3.10.4"
pytype = "*"
mypy = "*"
pyright = "*"
pyre-check = "*"
requests = "^2.31.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportMissingTypeStubs = false
reportUnknownParameterType = true
strict = ["."]

[tool.mypy]
strict = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_unreachable = true
plugins = "type_checker_features/mypy_directory/plugin.py"

[tool.ruff]
ignore = ["ALL"]